refactored seeds.rb
