+++ spec/db/seed_spec.rb
@@ -1,4 +1,5 @@
 require 'rails_helper'
+require 'rake'
 
 RSpec.describe 'load admin.email and admin.password from ENV in production' do
